Find SCCT port number + add debug logging
It was previously assumed that the scct websocket port was static. It is
actually a different port per profile. Profiles are stored in:
AppData/Local/team pheeniX/StarCraft-Casting-Tool/profiles
Each directory is a hex number that is also the websocket port number.
To find the active profile, a port scan is run over all the profile
ports to find if any are open.